refactor: modernize routing system to use ai complexity analyzer
- Delete adaptive_routing.py removing rule-based complexity scoring
- Update all imports to use new complexity_types and ai_complexity_analyzer
- Make route_thought methods async to support AI calls
- Update server_core and processors for async routing
- Fix test files to support new structure
