+++ tests/unit/test_complexity_metrics.py
@@ -1,7 +1,7 @@
 """TDD tests for ComplexityMetrics.complexity_score property calculation."""
 
 import pytest
-from src.mcp_server_mas_sequential_thinking.adaptive_routing import ComplexityMetrics
+from src.mcp_server_mas_sequential_thinking.routing.complexity_types import ComplexityMetrics
 
 
 class TestComplexityMetricsScoreCalculation:
